==> R/02_operations_geomatiques_elementaires.qmd <==
---
title: "Faire de la géomatique avec R"
author: "Antoine Peris"
format: html
editor: visual
lang: fr
---

Chargement du fichier.

```{r}
#| message: false
#| warning: false
library(readr)
library(dplyr)

perils_geo <- read_csv2("../data/perils_geo.csv")
```

Il se peut qu'un immeuble ait fait l'objet d'un péril ordinaire et d'un péril imminent. On va donc supperimer les duplicats des adresses.

```{r}
perils_geo <- perils_geo %>% filter(!duplicated(adress))
```

### Transformation en objet spatial

On charge d’abord les packages nécessaires :

-   `sf` pour manipuler des données spatiales,

-   `mapview` pour visualiser les points sur une carte interactive.

```{r}
#| message: false
#| warning: false
library(sf) 
library(mapview)
```

On transforme ensuite le tableau en un objet spatial (`sf`) en précisant les colonnes de coordonnées et le système de projection (ici WGS84, EPSG:4326).

```{r}
perils_geo_sf <- st_as_sf(perils_geo, coords = c("longitude", "latitude")) %>% 
  st_set_crs(4326)
```

Enfin, on visualise les points sur une carte interactive, colorés selon l’année.

```{r}
mapview(perils_geo_sf, zcol = "annee", cex = 4)
```

On sauvegarde notre nouvelle couche spatiale.

```{r}
#| message: false
#| warning: false
st_write(perils_geo_sf, "../data/perils_geo.gpkg", append = F)
```

### Opérations spatiales élémentaires

On va ensuite dénombrer les périls par IRIS. Pour cela, nous devons faire une intersection spatiale.

Chargements de données sur les IRIS de Bordeaux.

```{r}
#| message: false
#| warning: false
iris <- st_read("../data/se_iri24_s.shp")
```

Reprojection de nos deux jeux de données.

```{r}
iris <- st_transform(iris, 2154)
perils_geo_sf <- st_transform(perils_geo_sf, 2154)
```

On réalise une jointure spatiale entre notre couche sur les périls et celle des iris en vue d'un dénombrement.

```{r}
perils_geo_sf_iris <- st_join(perils_geo_sf, iris)
```

On effectue le dénombrement grâce à la combinaison de `group_by()` et de `summarise()`

```{r}
#| message: false
#| warning: false
nb_perils_iris <- perils_geo_sf_iris %>% 
  as_tibble() %>% 
  group_by(nom_iris, ident) %>% 
  summarise(n=n())
```

On ajoute la variable de dénombrement aux IRIS grâce à une jointure attributaire.

```{r}
iris <- left_join(iris, nb_perils_iris, by="ident")
```

Dans l'optique de faire une carte en cercles proportionnels, on effectue l'opération `st_point_on_surface()`. Pour de la cartographie, cette option est plus robuste que `st_centroid()` qui peut générer des points hors des polygones (dans le cas d'une forme en croissant par exemple).

```{r}
iris_pt <- st_point_on_surface(iris)
```

On effectue finalement une cartographie simple.

```{r}
#| message: false
#| warning: false
library(ggplot2)

ggplot()+
  geom_sf(data = iris)+
  geom_sf(data=iris_pt, aes(size=n), alpha=.6)+
  scale_size_area(max_size = 5)+
  coord_sf(datum = NA)+
  theme_bw()+
  labs(title = "Les arrêtés de périls d'immeubles à Bordeaux", 
       subtitle = "Période 2020-2025",
       size="Nombre\nd'arrêtés")
```
